Use rails_helper for workaround
Putting the workaround in the spec_helper results in a NameError as described in issue #6. The reason for this is that the spec_helper is loaded before Rails gets loaded. An easy fix for this is to put the workaround not in the spec_helper, but in the rails_helper. This helper gets loaded after
Rails, which solved the problem.